app/requirements.txt
Django>=4.0.4,<4.1
djangorestframework>=3.13.1,<3.14
psycopg2-binary==2.9.6
drf-yasg==1.21.7 
coverage==7.3.3
